feat: move _MyTasksScreenExpansionTile to components for reuse
