+++ apps/tasks/lib/screens/main_screen_subscreens/my_tasks_screen.dart
@@ -10,6 +10,7 @@ import 'package:tasks/dataclasses/subtask.dart';
 import 'package:tasks/screens/settings_screen.dart';
 import 'package:helpwave_widget/loading.dart';
 
+import '../../components/TaskExpansionTile.dart';
 import '../../dataclasses/task.dart';
 
 /// The Screen for showing all [Task]'s the [User] has in the current [ ]
@@ -99,10 +100,6 @@ class _MyTasksScreenState extends State<MyTasksScreen> {
 
   @override
   Widget build(BuildContext context) {
-    Color upcomingColor = const Color(0xFF5D5FEF);
-    Color inProgressColor = const Color(0xFFC79345);
-    Color doneColor = const Color(0xFF7A977E);
-
     return Consumer(
       builder: (BuildContext context, ThemeModel themeNotifier, _) => Scaffold(
         appBar: AppBar(
@@ -139,17 +136,17 @@ class _MyTasksScreenState extends State<MyTasksScreen> {
                   spacing: distanceSmall,
                   // TODO change the color off the expansion tiles
                   children: [
-                    _MyTasksScreenExpansionTile(
+                    TaskExpansionTile(
                       tasks: todo,
                       color: upcomingColor,
                       title: context.localization!.upcoming,
                     ),
-                    _MyTasksScreenExpansionTile(
+                    TaskExpansionTile(
                       tasks: inProgress,
                       color: inProgressColor,
                       title: context.localization!.inProgress,
                     ),
-                    _MyTasksScreenExpansionTile(
+                    TaskExpansionTile(
                       tasks: done,
                       color: doneColor,
                       title: context.localization!.done,
@@ -165,41 +162,3 @@ class _MyTasksScreenState extends State<MyTasksScreen> {
   }
 }
 
-class _MyTasksScreenExpansionTile extends StatelessWidget {
-  final List<TaskWithPatient> tasks;
-  final Color color;
-  final double circleSize = 8;
-  final String title;
-
-  const _MyTasksScreenExpansionTile({required this.tasks, required this.color, required this.title});
-
-  @override
-  Widget build(BuildContext context) {
-    return ExpansionTile(
-      textColor: color,
-      collapsedTextColor: color,
-      initiallyExpanded: true,
-      leading: SizedBox(
-        width: circleSize,
-        child: Center(
-          child: Circle(
-            color: color,
-            diameter: circleSize,
-          ),
-        ),
-      ),
-      title: Text("${title} (${tasks.length})"),
-      children: tasks
-          .map(
-            (task) => TaskCard(
-              task: task,
-              margin: const EdgeInsets.symmetric(
-                horizontal: paddingSmall,
-                vertical: paddingTiny,
-              ),
-            ),
-          )
-          .toList(),
-    );
-  }
-}
